requirements.txt
pytest==6.2.3
allure-pytest==2.8.40
allure-python-commons==2.8.40
pylint-fail-under
pytest-xdist
pytest-rerunfailures==9.1.1
httpx
requests-toolbelt==0.9.1
simplejson==3.17.2
ruamel.yaml==0.17.4
configparser==5.0.2
openpyxl
loguru
pytest-sugar
pathlib
jsonpath
PyYAML
pandas
demjson
pytest-parallel
hypothesis
